Rewrite plugin
This fixes an issue which would cause an infinite display of error message when a test failed.

This version of the code passes through the program output to the console when the quiet option is not passed, allowing console coloring to pass through as well.  When the quiet option is specified, the output is buffered, which, unfortunately, causes a loss of the console coloring.

This change also adds support for some more options, including the ability to specify the path to elm-test (it defaults to the elm-test installed in ./node_modules/.bin/, then the globally-installed elm-test), and the ability to pass along args to elm-test (such as the path to the elm-make compiler or the seed).

This change also removes the JSPM integration, only because I wasn't sure how to keep it in place with my changes, and wasn't sure what benefit it was providing.